requirements.txt
jupyterlab==4.0.5
lifelines==0.27.7
matplotlib==3.7.2
mlflow==2.9.2
numpy==1.23.5
pandas==2.0.3
plotly==5.18.0
scikit-learn>=0.24.0
scikit-survival==0.21.0
scipy==1.11.2
seaborn==0.13.0
python-dotenv==1.0.1
xgboost==2.0.3
kaleido==0.2.1